requirements.txt
art==5.9
python-dotenv==1.0.0
google-generativeai==0.4.1
consolemenu==1.0.1
console-menu==0.8.0
logging==0.4.9.6
matplotlib==3.10.1
numpy== 2.2.4